// backend/src/api/bitcoin/bitcoin-api.interface.ts
export namespace IBitcoinApi {
// ...
  export interface Block {
    hash: string; //  (string) the block hash (same as provided)
    confirmations: number; //  (numeric) The number of confirmations, or -1 if the block is not on the main chain
    size: number; //  (numeric) The block size
    height: number; //  (numeric) The block height or index
    version: number; //  (numeric) The block version
    versionHex: string; //  (string) The block version formatted in hexadecimal
    merkleroot: string; //  (string) The merkle root
    tx: Transaction[];
    time: number; //  (numeric) The block time expressed in UNIX epoch time
    mediantime: number; //  (numeric) The median block time expressed in UNIX epoch time
    nonce: number; //  (numeric) The nonce
    bits: string; //  (string) The bits
    difficulty: number; //  (numeric) The difficulty
    chainwork: string; //  (string) Expected number of hashes required to produce the chain up to this block (in hex)
    nTx: number; //  (numeric) The number of transactions in the block
    previousblockhash: string; //  (string) The hash of the previous block
    nextblockhash: string; //  (string) The hash of the next block
    ablastate?: AblaState; //  (object, optional) Adaptive Blocksize Limit Algorithm state in BCH
  }
// ...
  export interface VerboseTransaction {
    in_active_chain: boolean; //  (boolean) Whether specified block is in the active chain or not
    hex: string; //  (string) The serialized, hex-encoded data for 'txid'
    txid: string; //  (string) The transaction id (same as provided)
    hash: string; //  (string) The transaction hash
    size: number; //  (numeric) The serialized transaction size
    version: number; //  (numeric) The version
    locktime: number; //  (numeric) The lock time
    vin: VerboseVin[];
    vout: VerboseVout[];
    blockhash: string; //  (string) the block hash
    confirmations: number; //  (numeric) The confirmations
    blocktime: number; //  (numeric) The block time expressed in UNIX epoch time
    time: number; //  (numeric) Same as blocktime
    fee?: number; //  (numeric) The transaction fee in BCH, omitted if block undo data is not available
  }

  export interface VerboseBlock extends Block {
    tx: VerboseTransaction[]; // The transactions in the format of the getrawtransaction RPC. Different from verbosity = 1 "tx" result
  }
// ...
  export interface VerboseVin {
    txid?: string; //  (string) The transaction id
    vout?: number; //  (string)
    value?: number; //  (numeric) The value in BCH
    scriptSig?: {
      //  (json object) The script
      asm: string; //  (string) asm
      hex: string; //  (string) hex
      byteCodePattern?: {
        //  (json object)
        pattern: string; // (string) the pattern hex
        data: string[]; //  (array) the data hex
      };
      redeemScript?: {
        //  (json object)
        asm: string; //  (string) the asm
        hex: string; //  (string) the hex
        byteCodePattern?: {
          //  (json object)
          pattern: string; // (string) the pattern hex
          data: string[]; //  (array) the data hex
          p2shType: string; // (string) the p2sh type (either p2sh20 or p2sh32)
        };
      };
    };
    scriptPubKey: {
      //  (json object)
      asm: string; //  (string) the asm
      hex: string; //  (string) the hex
      type: string; //  (string) The type, eg. nonstandard, pubkey, pubkeyhash, scripthash, multisig, nulldata
      address?: string; //  (string) bitcoin address
      byteCodePattern?: {
        //  (json object)
        pattern: string; // (string) the pattern hex
        data: string[]; //  (array) the data hex
      };
    };
    tokenData?: {
      //  (json object)
      category: string; // token id
      amount: number; // fungible amount
      // (optional json object)
      nft?: {
        capability: string; // one of "none", "mutable", "minting"
        commitment: string; // NFT commitment
      };
    };
    sequence: number; //  (numeric) The script sequence number
    coinbase?: string; // (string) only present if coinbase transaction in hex
  }

  export interface VerboseVout {
    value: number; //  (numeric) The value in BCH
    n: number; //  (numeric) index
    scriptPubKey: {
      //  (json object)
      asm: string; //  (string) the asm
      hex: string; //  (string) the hex
      reqSigs?: number; //  (numeric) The required sigs
      type: string; //  (string) The type, eg. 'pubkeyhash'
      addresses?: string[]; //  (string) bitcoin addresses
      byteCodePattern?: {
        //  (json object)
        pattern: string; // (string) the pattern hex
        data: string[]; //  (array) the data hex
      };
    };
    tokenData?: {
      //  (json object)
      category: string; // token id
      amount: number; // fungible amount
      // (optional json object)
      nft?: {
        capability: string; // one of "none", "mutable", "minting"
        commitment: string; // NFT commitment
      };
    };
  }
// ...
}