// backend/src/api/bitcoin/public-api.interface.ts
export namespace IPublicApi {
// ...
  export interface VerboseTransaction {
    txid: string;
    version: number;
    locktime: number;
    size: number;
    fee: number;
    sigops?: number;
    vin: VerboseVin[];
    vout: VerboseVout[];
    status: Status;
    hex?: string;
    // TODO: Also add blockhash, confirmations, time and blocktime.. No need for this status object
  }
// ...
  export interface VerboseVin {
    txid: string;
    vout: number;
    value: number | null;
    is_coinbase: boolean;
    scriptsig: string; // in hex
    scriptsig_asm: string; // in asm
    inner_redeemscript_asm: string;
    scriptsig_byte_code_pattern: string; // in hex
    scriptsig_byte_code: string[]; // script data in hex
    scriptpubkey: string; // in hex
    scriptpubkey_asm: string;
    scriptpubkey_type: string;
    scriptpubkey_byte_code_pattern: string; // in hex
    scriptpubkey_byte_code: string[]; // script data in hex
    scriptpubkey_address?: string;
    token_category?: string;
    token_amount?: string;
    token_nft_capability?: string; // "none", "mutable", "minting"
    token_nft_commitment?: string; // in hex
    sequence: any;
    prevout: VerboseVout | null;
    // Custom
    lazy?: boolean;
  }

  export interface VerboseVout {
    scriptpubkey: string; // in hex
    scriptpubkey_asm: string;
    scriptpubkey_type: string;
    scriptpubkey_address?: string;
    scriptpubkey_byte_code_pattern: string; // in hex
    scriptpubkey_byte_code: string[]; // script data in hex
    token_category?: string;
    token_amount?: string;
    token_nft_capability?: string; // "none", "mutable", "minting"
    token_nft_commitment?: string; // in hex
    value: number;
  }

  export interface Status {
    confirmed: boolean;
    block_height?: number;
    block_hash?: string;
    block_time?: number;
  }
// ...
}